+++ src/main/java/radon/jujutsu_kaisen/ability/ten_shadows/ShadowStorage.java
@@ -11,6 +11,8 @@ import radon.jujutsu_kaisen.ability.base.Ability;
 import radon.jujutsu_kaisen.ability.MenuType;
 import radon.jujutsu_kaisen.capability.data.ISorcererData;
 import radon.jujutsu_kaisen.capability.data.SorcererDataHandler;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.ITenShadowsData;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.TenShadowsDataHandler;
 import radon.jujutsu_kaisen.client.ClientWrapper;
 import radon.jujutsu_kaisen.util.HelperMethods;
 import radon.jujutsu_kaisen.util.RotationUtil;
@@ -50,10 +52,9 @@ public class ShadowStorage extends Ability {
         if (owner.isShiftKeyDown()) {
             if (owner.getMainHandItem().isEmpty()) return;
 
-            owner.getCapability(SorcererDataHandler.INSTANCE).ifPresent(cap -> {
-                cap.addShadowInventory(owner.getMainHandItem());
-                owner.setItemSlot(EquipmentSlot.MAINHAND, ItemStack.EMPTY);
-            });
+            ITenShadowsData cap = owner.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
+            cap.addShadowInventory(owner.getMainHandItem());
+            owner.setItemSlot(EquipmentSlot.MAINHAND, ItemStack.EMPTY);
         } else if (owner.level().isClientSide) {
             ClientWrapper.openShadowInventory();
         }
@@ -61,13 +62,12 @@ public class ShadowStorage extends Ability {
 
     @Override
     public Status isTriggerable(LivingEntity owner) {
-        if (!owner.getCapability(SorcererDataHandler.INSTANCE).isPresent()) return Status.FAILURE;
-        ISorcererData cap = owner.getCapability(SorcererDataHandler.INSTANCE).resolve().orElseThrow();
+        ITenShadowsData cap = owner.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
 
         if (owner.isShiftKeyDown()) {
             if (owner.getMainHandItem().isEmpty()) return Status.FAILURE;
         } else {
-            if (cap.getShadowInventory().size() == 0) return Status.FAILURE;
+            if (cap.getShadowInventory().isEmpty()) return Status.FAILURE;
         }
         return super.isTriggerable(owner);
     }
