+++ src/main/java/radon/jujutsu_kaisen/ability/ten_shadows/summon/DivineDogs.java
@@ -12,6 +12,8 @@ import radon.jujutsu_kaisen.ability.JJKAbilities;
 import radon.jujutsu_kaisen.ability.base.Summon;
 import radon.jujutsu_kaisen.capability.data.ISorcererData;
 import radon.jujutsu_kaisen.capability.data.SorcererDataHandler;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.ITenShadowsData;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.TenShadowsDataHandler;
 import radon.jujutsu_kaisen.entity.JJKEntities;
 import radon.jujutsu_kaisen.entity.ten_shadows.*;
 import radon.jujutsu_kaisen.network.PacketHandler;
@@ -57,7 +59,7 @@ public class DivineDogs extends Summon<DivineDogEntity> {
     @Override
     protected boolean isDead(LivingEntity owner, EntityType<?> type) {
         if (!owner.getCapability(SorcererDataHandler.INSTANCE).isPresent()) return false;
-        ISorcererData cap = owner.getCapability(SorcererDataHandler.INSTANCE).resolve().orElseThrow();
+        ITenShadowsData cap = owner.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
         Registry<EntityType<?>> registry = owner.level().registryAccess().registryOrThrow(Registries.ENTITY_TYPE);
         return cap.isDead(registry, JJKEntities.DIVINE_DOG_WHITE.get()) && cap.isDead(registry, JJKEntities.DIVINE_DOG_BLACK.get());
     }
@@ -65,21 +67,22 @@ public class DivineDogs extends Summon<DivineDogEntity> {
     @Override
     public void spawn(LivingEntity owner, boolean clone) {
         if (!owner.level().isClientSide) {
-            ISorcererData cap = owner.getCapability(SorcererDataHandler.INSTANCE).resolve().orElseThrow();
+            ITenShadowsData tenShadowsCap = owner.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
+            ISorcererData sorcererCap = owner.getCapability(SorcererDataHandler.INSTANCE).resolve().orElseThrow();
 
             Registry<EntityType<?>> registry = owner.level().registryAccess().registryOrThrow(Registries.ENTITY_TYPE);
 
-            if (!cap.isDead(registry, JJKEntities.DIVINE_DOG_WHITE.get())) {
+            if (!tenShadowsCap.isDead(registry, JJKEntities.DIVINE_DOG_WHITE.get())) {
                 DivineDogWhiteEntity white = new DivineDogWhiteEntity(owner, false);
                 white.setClone(clone);
                 owner.level().addFreshEntity(white);
-                cap.addSummon(white);
+                sorcererCap.addSummon(white);
             }
-            if (!cap.isDead(registry, JJKEntities.DIVINE_DOG_BLACK.get())) {
+            if (!tenShadowsCap.isDead(registry, JJKEntities.DIVINE_DOG_BLACK.get())) {
                 DivineDogBlackEntity black = new DivineDogBlackEntity(owner, false);
                 black.setClone(clone);
                 owner.level().addFreshEntity(black);
-                cap.addSummon(black);
+                sorcererCap.addSummon(black);
             }
         }
     }
