+++ src/main/java/radon/jujutsu_kaisen/capability/data/ISorcererData.java
@@ -255,58 +255,6 @@ public interface ISorcererData {
 
     <T extends Entity> boolean hasSummonOfClass(Class<T> clazz);
 
-    boolean hasTamed(Registry<EntityType<?>> registry, EntityType<?> entity);
-    void tame(Registry<EntityType<?>> registry, EntityType<?> entity);
-    void setTamed(Set<ResourceLocation> tamed);
-    Set<ResourceLocation> getTamed();
-    boolean isDead(Registry<EntityType<?>> registry, EntityType<?> entity);
-    Set<ResourceLocation> getDead();
-    void setDead(Set<ResourceLocation> dead);
-
-    void kill(Registry<EntityType<?>> registry, EntityType<?> entity);
-
-    void revive(boolean full);
-
-    Set<Adaptation> getAdapted();
-
-    void addAdapted(Set<Adaptation> adaptations);
-
-    Map<Adaptation, Integer> getAdapting();
-
-    void addAdapting(Map<Adaptation, Integer> adapting);
-
-    void addShadowInventory(ItemStack stack);
-
-    ItemStack getShadowInventory(int index);
-
-    List<ItemStack> getShadowInventory();
-
-    void removeShadowInventory(int index);
-
-    float getAdaptationProgress(DamageSource source);
-
-    float getAdaptationProgress(Adaptation adaptation);
-
-    Adaptation.Type getAdaptationType(DamageSource source);
-
-    Adaptation.Type getAdaptationType(Adaptation adaptation);
-
-    Map<Adaptation.Type, Float> getAdaptationTypes();
-
-    boolean isAdaptedTo(DamageSource source);
-
-    boolean isAdaptedTo(@Nullable Ability ability);
-
-    boolean isAdaptedTo(CursedTechnique technique);
-
-    void tryAdapt(DamageSource source);
-
-    void tryAdapt(Ability ability);
-
-    TenShadowsMode getMode();
-
-    void setMode(TenShadowsMode mode);
-
     void addCurse(AbsorbedCurse curse);
 
     void removeCurse(AbsorbedCurse curse);
