+++ src/main/java/radon/jujutsu_kaisen/capability/data/ten_shadows/TenShadowsDataHandler.java
@@ -1,9 +1,8 @@
-package radon.jujutsu_kaisen.capability.data;
+package radon.jujutsu_kaisen.capability.data.ten_shadows;
 
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
@@ -20,15 +19,14 @@ import net.minecraftforge.fml.common.Mod;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import radon.jujutsu_kaisen.JujutsuKaisen;
-import radon.jujutsu_kaisen.capability.data.ten_shadows.TenShadowsDataHandler;
+import radon.jujutsu_kaisen.capability.data.ISorcererData;
+import radon.jujutsu_kaisen.capability.data.SorcererDataHandler;
 import radon.jujutsu_kaisen.config.ConfigHolder;
 import radon.jujutsu_kaisen.entity.base.ISorcerer;
-import radon.jujutsu_kaisen.network.PacketHandler;
-import radon.jujutsu_kaisen.network.packet.s2c.SyncSorcererDataS2CPacket;
 
 @Mod.EventBusSubscriber(modid = JujutsuKaisen.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
-public class SorcererDataHandler {
-    public static Capability<ISorcererData> INSTANCE = CapabilityManager.get(new CapabilityToken<>() {});
+public class TenShadowsDataHandler {
+    public static Capability<ITenShadowsData> INSTANCE = CapabilityManager.get(new CapabilityToken<>() {});
 
     @SubscribeEvent
     public static void onPlayerClone(PlayerEvent.Clone event) {
@@ -37,23 +35,14 @@ public class SorcererDataHandler {
 
         original.reviveCaps();
 
-        ISorcererData oldCap = original.getCapability(INSTANCE).resolve().orElseThrow();
-        ISorcererData newCap = player.getCapability(INSTANCE).resolve().orElseThrow();
+        ITenShadowsData oldCap = original.getCapability(INSTANCE).resolve().orElseThrow();
+        ITenShadowsData newCap = player.getCapability(INSTANCE).resolve().orElseThrow();
 
         newCap.deserializeNBT(oldCap.serializeNBT());
 
         if (event.isWasDeath()) {
-            newCap.setEnergy(newCap.getMaxEnergy());
-            newCap.resetCooldowns();
-            newCap.resetBurnout();
-            newCap.clearToggled();
-            
-            newCap.resetBlackFlash();
-            newCap.resetExtraEnergy();
-            newCap.resetSpeedStacks();
-
-            if (!player.level().isClientSide) {
-                PacketHandler.sendToClient(new SyncSorcererDataS2CPacket(newCap.serializeNBT()), (ServerPlayer) player);
+            if (!ConfigHolder.SERVER.realisticShikigami.get()) {
+                newCap.revive(false);
             }
         }
         original.invalidateCaps();
@@ -63,23 +52,23 @@ public class SorcererDataHandler {
     public static void onAttachCapabilities(AttachCapabilitiesEvent<Entity> event) {
         if (event.getObject() instanceof LivingEntity entity) {
             if (entity instanceof Player || entity instanceof ISorcerer) {
-                SorcererDataProvider provider = new SorcererDataProvider();
-                ISorcererData cap = provider.getCapability(INSTANCE).resolve().orElseThrow();
+                TenShadowsDataProvider provider = new TenShadowsDataProvider();
+                ITenShadowsData cap = provider.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
                 cap.init(entity);
-                event.addCapability(SorcererDataProvider.IDENTIFIER, provider);
+                event.addCapability(TenShadowsDataProvider.IDENTIFIER, provider);
             }
         }
     }
+    
+    public static class TenShadowsDataProvider implements ICapabilityProvider, INBTSerializable<CompoundTag> {
+        public static ResourceLocation IDENTIFIER = new ResourceLocation(JujutsuKaisen.MOD_ID, "ten_shadows_data");
 
-    public static class SorcererDataProvider implements ICapabilityProvider, INBTSerializable<CompoundTag> {
-        public static ResourceLocation IDENTIFIER = new ResourceLocation(JujutsuKaisen.MOD_ID, "sorcerer_data");
+        private ITenShadowsData cap = null;
+        private final LazyOptional<ITenShadowsData> optional = LazyOptional.of(this::create);
 
-        private ISorcererData cap = null;
-        private final LazyOptional<ISorcererData> optional = LazyOptional.of(this::create);
-
-        private ISorcererData create() {
+        private ITenShadowsData create() {
             if (this.cap == null) {
-                this.cap = new SorcererData();
+                this.cap = new TenShadowsData();
             }
             return this.cap;
         }
