+++ src/main/java/radon/jujutsu_kaisen/network/packet/c2s/ShadowInventoryTakeC2SPacket.java
@@ -5,9 +5,13 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.network.NetworkEvent;
+import radon.jujutsu_kaisen.capability.data.ISorcererData;
 import radon.jujutsu_kaisen.capability.data.SorcererDataHandler;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.ITenShadowsData;
+import radon.jujutsu_kaisen.capability.data.ten_shadows.TenShadowsDataHandler;
 import radon.jujutsu_kaisen.network.PacketHandler;
 import radon.jujutsu_kaisen.network.packet.s2c.SyncSorcererDataS2CPacket;
+import radon.jujutsu_kaisen.network.packet.s2c.SyncTenShadowsDataS2CPacket;
 
 import java.util.function.Supplier;
 
@@ -34,18 +38,18 @@ public class ShadowInventoryTakeC2SPacket {
 
             assert sender != null;
 
-            sender.getCapability(SorcererDataHandler.INSTANCE).ifPresent(cap -> {
-                ItemStack stack = cap.getShadowInventory(this.index);
+            ITenShadowsData cap = sender.getCapability(TenShadowsDataHandler.INSTANCE).resolve().orElseThrow();
 
-                if (sender.getMainHandItem().isEmpty()) {
-                    sender.setItemSlot(EquipmentSlot.MAINHAND, stack);
-                } else {
-                    if (!sender.addItem(stack)) return;
-                }
-                cap.removeShadowInventory(this.index);
+            ItemStack stack = cap.getShadowInventory(this.index);
 
-                PacketHandler.sendToClient(new SyncSorcererDataS2CPacket(cap.serializeNBT()), sender);
-            });
+            if (sender.getMainHandItem().isEmpty()) {
+                sender.setItemSlot(EquipmentSlot.MAINHAND, stack);
+            } else {
+                if (!sender.addItem(stack)) return;
+            }
+            cap.removeShadowInventory(this.index);
+
+            PacketHandler.sendToClient(new SyncTenShadowsDataS2CPacket(cap.serializeNBT()), sender);
         });
         ctx.setPacketHandled(true);
     }
